requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic-settings
alembic
pytest
httpx
python-dotenv